requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
whitenoise[brotli]>=6.6.0
